+++ src/main/java/com/example/onlineBriefing/repositories/ProfileBriefingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfileBriefingRepository extends JpaRepository<ProfileBriefing,Long> {
+public interface ProfileBriefingRepository extends JpaRepository<ProfileBriefing,Integer> {
 }
