+++ src/main/java/com/example/onlineBriefing/repositories/TeacherRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeacherRepository extends JpaRepository<Teacher,Long> {
+public interface TeacherRepository extends JpaRepository<Teacher,Integer> {
     Teacher findByLogin(Integer login);
 }
