+++ src/main/java/com/example/onlineBriefing/repositories/UniversityGroupRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UniversityGroupRepository extends JpaRepository<UniversityGroup,Long> {
+public interface UniversityGroupRepository extends JpaRepository<UniversityGroup,Integer> {
 }
